mason_try_binary: fixup installed .la and .pc files
Previously only MASON_PKGCONFIG_FILE was amended, and only partially.
Absolute paths that were not on "prefix=" lines were left untouched.

This commit adds MASON_ROOT to mason.ini, because that's the path that
should be replaced and there was no way to obtain its build-time value
during installation.

The build-time value of MASON_ROOT is then replaced everywhere in any
libtool and pkgconfig files found.